storage: Move MergeInternalTimeSeriesData to use Go-based MVCCMerger
Currently, some timeseries merge tests call into
storage.MergeInternalTimeSeriesData, which used to use the merger
written in C++. As part of the move to delete all RocksDB-related
C++ code, that method needs to be moved to use the Go-based
MVCCMerger. This change does just that.

Also delete the mergeIntoNil value, as the difference doesn't
matter in the case of MVCCMerger.

Release note: None.
