+++ pkg/storage/rocksdb.go
@@ -2506,19 +2506,6 @@ func goMerge(existing, update []byte) ([]byte, error) {
 	return cStringToGoBytes(result), nil
 }
 
-// goPartialMerge takes existing and update byte slices that are expected to
-// be marshaled roachpb.Values and performs a partial merge using C++ code,
-// marshaled roachpb.Value or an error.
-func goPartialMerge(existing, update []byte) ([]byte, error) {
-	var result C.DBString
-	status := C.DBPartialMergeOne(goToCSlice(existing), goToCSlice(update), &result)
-	if status.data != nil {
-		return nil, errors.Errorf("%s: existing=%q, update=%q",
-			cStringToGoString(status), existing, update)
-	}
-	return cStringToGoBytes(result), nil
-}
-
 func dbPut(rdb *C.DBEngine, key MVCCKey, value []byte) error {
 	if len(key.Key) == 0 {
 		return emptyKeyError()
